requirements.txt
gpt4all==2.8.2
sentence-transformers==2.7.0
pypdf==4.3.1
tqdm==4.66.4
streamlit==1.39.0
llama-cpp-python
scikit-learn==1.7.2
numpy==2.3.3